requirements.txt
tensorflow==2.18.0
keras
numpy
matplotlib
scikit-learn
flask
pillow
